Change about page to organizers and update all  imports. Also make links more robust.
