--- src/about/teamNames.js
@@ -1,125 +1,0 @@
-import Joseph_Theoharakis from '../img/headshots/Joseph_Theoharakis.png'
-import Celine_Khoo from '../img/headshots/Celine_Khoo.jpg'
-import Denise_Weinberger from '../img/headshots/Denise_Weinberger.jpg'
-import Matt_Rantz from '../img/headshots/Matt_Rantz.jpg'
-
-import Mridul_Awasthi from '../img/headshots/Mridul_Awasthi.jpg'
-import Saem_Jeon from '../img/headshots/Saem_Jeon.jpg'
-import Atandrila_Anuva from '../img/headshots/Atandrila_Anuva.jpg'
-
-import Jason_Carrete from '../img/headshots/Jason_Carrete.jpg'
-import Zach_Matuson from '../img/headshots/Zach_Matuson.jpg'
-import Hoang_Nguyen from '../img/headshots/Hoang_Nguyen.png'
-import Steven_Kearns from '../img/headshots/Steven_Kearns.png'
-
-import Daniel_Erbynn from '../img/headshots/Daniel_Erbynn.jpg'
-import Har_Patel from '../img/headshots/Har_Patel.png'
-import Madeline_Harmes from '../img/headshots/Madeline_Harmes.jpg'
-import Tapasya_Sharma from '../img/headshots/Tapasya_Sharma.jpg'
-
-import Alisha_Augustin from '../img/headshots/Alisha_Augustin.jpg'
-
-import Avi_Majumdar from '../img/headshots/Avi_Majumdar.png'
-import Julia_Fleming from '../img/headshots/Julia_Fleming.jpg'
-import Vincent_Theodor_Go from '../img/headshots/Vincent_Theodor_Go.jpg'
-
-export const execBoard = [
-  {
-    imgSrc: Joseph_Theoharakis,
-    name: 'Joseph Theoharakis',
-    title: 'President'
-  },
-  {
-    imgSrc: Celine_Khoo,
-    name: 'Celine Khoo',
-    title: 'Vice President'
-  },
-  {
-    imgSrc: Denise_Weinberger,
-    name: 'Denise Weinberger',
-    title: 'Treasurer'
-  },
-  {
-    imgSrc: Matt_Rantz,
-    name: 'Matt Rantz',
-    title: 'Event Coordinator'
-  }
-]
-
-export const coordinators = [
-  {
-    imgSrc: Mridul_Awasthi,
-    name: 'Mridul Awasthi'
-  },
-  {
-    imgSrc: Saem_Jeon,
-    name: 'Saem Jeon'
-  },
-  {
-    imgSrc: Atandrila_Anuva,
-    name: 'Atandrila Anuva'
-  }
-]
-
-export const websiteDevelopers = [
-  {
-    imgSrc: Jason_Carrete,
-    name: 'Jason Carrete',
-    title: 'Website Master'
-  },
-  {
-    imgSrc: Zach_Matuson,
-    name: 'Zach Matuson'
-  },
-  {
-    imgSrc: Hoang_Nguyen,
-    name: 'Hoang Nguyen'
-  },
-  {
-    imgSrc: Steven_Kearns,
-    name: 'Steven Kearns'
-  }
-]
-
-export const sponsorshipTeam = [
-  {
-    imgSrc: Daniel_Erbynn,
-    name: 'Daniel Erbynn'
-  },
-  {
-    imgSrc: Har_Patel,
-    name: 'Har Patel'
-  },
-  {
-    imgSrc: Madeline_Harmes,
-    name: 'Madeline Harmes'
-  },
-  {
-    imgSrc: Tapasya_Sharma,
-    name: 'Tapasya Sharma'
-  }
-]
-
-export const designTeam = [
-  // Shara
-  // Kaley Nhu
-  {
-    imgSrc: Alisha_Augustin,
-    name: 'Alisha Augustin'
-  }
-]
-
-export const marketingTeam = [
-  {
-    imgSrc: Avi_Majumdar,
-    name: 'Avi Majumdar'
-  },
-  {
-    imgSrc: Julia_Fleming,
-    name: 'Julia Fleming'
-  },
-  {
-    imgSrc: Vincent_Theodor_Go,
-    name: 'Vincent Theodor Go'
-  }
-]
