+++ src/teamNames.js
@@ -3,21 +3,27 @@ import Celine_Khoo from './img/headshots/Celine_Khoo.jpg'
 import Denise_Weinberger from './img/headshots/Denise_Weinberger.jpg'
 import Matt_Rantz from './img/headshots/Matt_Rantz.jpg'
 
+import Mridul_Awasthi from './img/headshots/Mridul_Awasthi.jpg'
+import Saem_Jeon from './img/headshots/Saem_Jeon.jpg'
+import Atandrila_Anuva from './img/headshots/Atandrila_Anuva.jpg'
+
 import Jason_Carrete from './img/headshots/Jason_Carrete.jpg'
 import Zach_Matuson from './img/headshots/Zach_Matuson.jpg'
 import Hoang_Nguyen from './img/headshots/Hoang_Nguyen.png'
 import Steven_Kearns from './img/headshots/Steven_Kearns.png'
-import Alisha_Augustin from './img/headshots/Alisha_Augustin.jpg'
-import Atandrila_Anuva from './img/headshots/Atandrila_Anuva.jpg'
-import Julia_Fleming from './img/headshots/Julia_Fleming.jpg'
-import Mridul_Awasthi from './img/headshots/Mridul_Awasthi.jpg'
-import Daniel_Erbynn from './img/headshots/Daniel_Erbynn.jpg'
-import Vincent_Theodor_Go from './img/headshots/Vincent_Theodor_Go.jpg'
-import Avi_Majumdar from './img/headshots/Avi_Majumdar.png'
-import Tapasya_Sharma from './img/headshots/Tapasya_Sharma.jpg'
-import Saem_Jeon from './img/headshots/Saem_Jeon.jpg'
 
-export const execBoardData = [
+import Daniel_Erbynn from './img/headshots/Daniel_Erbynn.jpg'
+import Har_Patel from './img/headshots/Har_Patel.png'
+import Madeline_Harmes from './img/headshots/Madeline_Harmes.jpg'
+import Tapasya_Sharma from './img/headshots/Tapasya_Sharma.jpg'
+
+import Alisha_Augustin from './img/headshots/Alisha_Augustin.jpg'
+
+import Avi_Majumdar from './img/headshots/Avi_Majumdar.png'
+import Julia_Fleming from './img/headshots/Julia_Fleming.jpg'
+import Vincent_Theodor_Go from './img/headshots/Vincent_Theodor_Go.jpg'
+
+export const execBoard = [
   {
     imgSrc: Joseph_Theoharakis,
     name: 'Joseph Theoharakis',
@@ -40,7 +46,22 @@ export const execBoardData = [
   }
 ]
 
-export const websiteMasterData = [
+export const coordinators = [
+  {
+    imgSrc: Mridul_Awasthi,
+    name: 'Mridul Awasthi'
+  },
+  {
+    imgSrc: Saem_Jeon,
+    name: 'Saem Jeon'
+  },
+  {
+    imgSrc: Atandrila_Anuva,
+    name: 'Atandrila Anuva'
+  }
+]
+
+export const websiteDevelopers = [
   {
     imgSrc: Jason_Carrete,
     name: 'Jason Carrete',
@@ -59,35 +80,39 @@ export const websiteMasterData = [
     name: 'Steven Kearns'
   }
 ]
-/*
-Website masters who have not sent their pictures
 
-Zach Matusen
-*/
-
-export const committeeData = [
-  {
-    imgSrc: Mridul_Awasthi,
-    name: 'Mridul Awasthi',
-    title: 'DragonHacks Coordinator'
-  },
-  {
-    imgSrc: Atandrila_Anuva,
-    name: 'Atandrila Anuva',
-    title: 'DragonHacks Coordinator'
-  },
-  {
-    imgSrc: Saem_Jeon,
-    name: 'Saem Jeon',
-    title: 'DragonHacks Coordinator'
-  },
-  {
-    imgSrc: Alisha_Augustin,
-    name: 'Alisha Augustin'
-  },
+export const sponsorshipTeam = [
   {
     imgSrc: Daniel_Erbynn,
     name: 'Daniel Erbynn'
+  },
+  {
+    imgSrc: Har_Patel,
+    name: 'Har Patel'
+  },
+  {
+    imgSrc: Madeline_Harmes,
+    name: 'Madeline Harmes'
+  },
+  {
+    imgSrc: Tapasya_Sharma,
+    name: 'Tapasya Sharma'
+  }
+]
+
+export const designTeam = [
+  // Shara
+  // Kaley Nhu
+  {
+    imgSrc: Alisha_Augustin,
+    name: 'Alisha Augustin'
+  }
+]
+
+export const marketingTeam = [
+  {
+    imgSrc: Avi_Majumdar,
+    name: 'Avi Majumdar'
   },
   {
     imgSrc: Julia_Fleming,
@@ -96,25 +121,5 @@ export const committeeData = [
   {
     imgSrc: Vincent_Theodor_Go,
     name: 'Vincent Theodor Go'
-  },
-  {
-    imgSrc: Avi_Majumdar,
-    name: 'Avi Majumdar'
-  },
-  {
-    imgSrc: Tapasya_Sharma,
-    name: 'Tapasya Sharma'
   }
 ]
-
-/*
-Committee members who have not sent their pictures:
-
-Aviroop Majumdar
-Gabby Dela Paz
-Har Patel
-Kristopher Lopez
-Steven Lee
-Viwing Zheng
-Yigit Alparslan
-*/
